Switch to go-fuse RawFilesystem API (#173)
This switches onedriver to the go-fuse RawFilesystem API. The old filesystem setup had an extra "fileystem" it was using from go-fuse on top of the one we already wrote because that API was easier to use. By removing this extra layer, it cuts out a layer of abstraction and makes the codebase make a lot more intuitive (now there's an actual filesystem+inodes instead of a cache+"fs-like inodes").